requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.3.5
PyMuPDF>=1.19.0
pydantic>=1.9.0
tqdm>=4.62.3

# Embedding models and LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0
qdrant-client>=1.15.0
tiktoken>=0.5.0

# PDF Processing (Layout Aware)
docling>=1.0.0

# Web interaction
playwright>=1.40.0
beautifulsoup4>=4.12.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.5
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
